Global chat structure fix
