+++ ari/__main__.py
@@ -93,7 +93,7 @@ def main():
     try:
         loop = asyncio.new_event_loop()
         asyncio.set_event_loop(loop)
-
+        
         # Loads all necessary configurations for the bot
         data_mananger.load_basic_configuration()
