--- ari/plugins/ui.py
@@ -1,52 +1,0 @@
-
-from typing import List
-import discord
-from discord.utils import MISSING
-
-class LobbyDropDown(discord.ui.Select):
-    def __init__(self,server_lobbies,author, on_item_added):
-        self.server_lobbies = server_lobbies
-        self.author = author
-        self.on_item_added = on_item_added
-        
-        options = [discord.SelectOption(label=lobby["lobbyname"], value=lobby["lobbyname"]) for lobby in self.server_lobbies]
-        super().__init__(
-            placeholder="Select a lobby",
-            options=options,
-            min_values=1,
-            max_values=1
-        )
-    async def callback(self, interaction):
-        if interaction.user == self.author:
-            await interaction.response.defer()
-            await self.on_item_added(interaction.data['values'][0])
-            
-        
-class ConnectDropDown(discord.ui.View):
-    def __init__(self, author, server_lobbies):
-        super().__init__()
-        self.lobby = None
-        self.add_item(LobbyDropDown(server_lobbies,author, self.on_item_added))
-
-    async def on_item_added(self,value):
-        self.lobby = value
-        self.stop()
-class Choice(discord.ui.View):
-    def __init__(self, author):
-        super().__init__()
-        self.author = author
-        self.value = None
-
-    @discord.ui.button(label="Yes" , style=discord.ButtonStyle.green)
-    async def btn1(self, interaction: discord.interactions, btn:discord.ui.button):
-        if interaction.user == self.author:
-            self.value = True
-            await interaction.response.defer()
-            self.stop()
-
-    @discord.ui.button(label="Back" , style=discord.ButtonStyle.red)
-    async def btn2(self, interaction: discord.interactions, btn:discord.ui.button):
-        if interaction.user == self.author:
-            self.value = False
-            await interaction.response.defer()
-            self.stop()
